+++ examples/dogsvscats/model_inceptionv3.go
@@ -12,11 +12,14 @@ import (
 )
 
 // InceptionV3ModelPrep is executed before training: it downloads the inceptionv3 weights.
-func InceptionV3ModelPrep(scope *model.Scope, dataDir string, checkpointHandler *checkpoint.Handler) {
+func InceptionV3ModelPrep(scope *model.Scope, dataDir string, checkpointHandler *checkpoint.Handler) error {
 	scope.SetParam("data_dir", dataDir)
 	if model.GetParamOr(scope, "inception_pretrained", true) {
-		check(inceptionv3.DownloadAndUnpackWeights(dataDir))
+		if err := inceptionv3.DownloadAndUnpackWeights(dataDir); err != nil {
+			return err
+		}
 	}
+	return nil
 }
 
 // InceptionV3ModelGraph uses an optionally pre-trained inception model.
@@ -28,10 +31,8 @@ func InceptionV3ModelPrep(scope *model.Scope, dataDir string, checkpointHandler 
 // Results if we don't use the pre-trained weights (it can probably get much better with more training):
 // - no scaling (from 0.0 to 1.0): 62.5% accuracy
 // - with Keras scale (from -1.0 to 1.0): 61.8% accuracy
-func InceptionV3ModelGraph(scope *model.Scope, spec any, inputs []*Node) []*Node {
+func InceptionV3ModelGraph(scope *model.Scope, images *Node) *Node {
 	scope = scope.In("model") // Create the model by default under the "/model" scope.
-	_ = spec                  // Not needed.
-	images := inputs[0]       // Images scaled from 0.0 to 1.0
 	channelsConfig := timage.ChannelsLast
 	images = inceptionv3.PreprocessImage(images, 1.0, channelsConfig) // Adjust image to format used by Inception.
 	dataDir := model.GetParamOr(scope, "data_dir", ".")
@@ -46,5 +47,5 @@ func InceptionV3ModelGraph(scope *model.Scope, spec any, inputs []*Node) []*Node
 		Trainable(model.GetParamOr(scope, "inception_finetuning", false)).
 		Done()
 	logits = fnn.New(scope.In("fnn"), logits, 1).Done()
-	return []*Node{logits}
+	return logits
 }